+++ packages/fields/src/types/Float/filterTests.js
@@ -28,8 +28,14 @@ export const initItems = () => {
 };
 
 export const filterTests = withKeystone => {
-  const match = (keystone, filter, targets) =>
-    matchFilter(keystone, filter, '{ name, stars }', targets, 'name');
+  const match = (keystone, queryArgs, expected) =>
+    matchFilter({
+      keystone,
+      queryArgs,
+      fieldSelection: 'name stars',
+      expected,
+      sortKey: 'name',
+    });
 
   test(
     'No filter',