+++ test-projects/client-validation/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/cypress-project-client-validation
+
 # @keystone-alpha/cypress-project-client-validation
 
 ## 1.0.17